Temporary fix for annoying submodules
Adds ignore dirty status on submodules, this is always do to line
endings and I am not sure how to just fix the actual problem. Apparently
there is a patch in the works for this, but you'd have to be running the
latest git build which means GHFW doesn't have it.